feature: navigation to home page
