+++ app/hotel-details.tsx
